Вынести определение недели в отдельную функцию
