+++ notification_service/tools/__init__.py
@@ -1,6 +1,17 @@
 from datetime import datetime, timedelta
 
 
+def find_week():
+    """Определение текущей недели"""
+    now = datetime.now()
+    sep = datetime(now.year if now.month >= 9 else now.year - 1, 9, 1)
+    d1 = sep - timedelta(days=sep.weekday())
+    d2 = now - timedelta(days=now.weekday())
+
+    parity = ((d2 - d1).days // 7) % 2
+    return 'odd' if parity else 'even'
+
+
 def forming_user_to_submit(
         chat_id: int,
         group: str,
